flattened FRI crate structure
