+++ fri/Cargo.toml
@@ -14,7 +14,7 @@ edition = "2018"
 bench = false
 
 [[bench]]
-name = "monolith_prover"
+name = "prover"
 harness = false
 
 [[bench]]
