+++ fri/src/lib.rs
@@ -3,20 +3,17 @@
 // This source code is licensed under the MIT license found in the
 // LICENSE file in the root directory of this source tree.
 
-mod prover;
-pub use prover::{DefaultProverChannel, FriProver, ProverChannel};
-
-mod verifier;
-pub use verifier::{verify, DefaultVerifierChannel, VerifierChannel, VerifierContext};
+pub mod folding;
+pub mod prover;
+pub mod verifier;
 
 mod options;
 pub use options::FriOptions;
 
-mod errors;
-pub use errors::{ProofSerializationError, VerifierError};
-
 mod proof;
-pub use proof::{FriProof, FriProofLayer};
+pub use proof::FriProof;
 
-pub mod folding;
-pub mod utils;
+mod errors;
+pub use errors::VerifierError;
+
+mod utils;
