+++ fri/src/prover/tests.rs
@@ -3,15 +3,49 @@
 // This source code is licensed under the MIT license found in the
 // LICENSE file in the root directory of this source tree.
 
+use super::{DefaultProverChannel, FriProver};
 use crate::{
-    verifier, DefaultProverChannel, DefaultVerifierChannel, FriOptions, FriProof, VerifierContext,
-    VerifierError,
+    verifier::{self, DefaultVerifierChannel, VerifierContext},
+    FriOptions, FriProof, VerifierError,
 };
 use crypto::{hashers::Blake3_256, Hasher, PublicCoin};
 use math::{fft, fields::f128::BaseElement, FieldElement};
 use utils::{Deserializable, Serializable, SliceReader};
 
 type Blake3 = Blake3_256<BaseElement>;
+
+// PROVE/VERIFY TEST
+// ================================================================================================
+
+#[test]
+fn fri_prove_verify() {
+    let trace_length = 4096;
+    let lde_blowup = 8;
+
+    let options = FriOptions::new(lde_blowup, 4, 256);
+    let mut channel = build_prover_channel(trace_length, &options);
+    let evaluations = build_evaluations(trace_length, lde_blowup);
+
+    // instantiate the prover and generate the proof
+    let mut prover = FriProver::new(options.clone());
+    prover.build_layers(&mut channel, evaluations.clone());
+    let positions = channel.draw_query_positions();
+    let proof = prover.build_proof(&positions);
+
+    // make sure the proof can be verified
+    let commitments = channel.fri_layer_commitments().to_vec();
+    let max_degree = trace_length - 1;
+    let result = verify_proof(
+        proof,
+        commitments,
+        &evaluations,
+        max_degree,
+        trace_length * lde_blowup,
+        &positions,
+        &options,
+    );
+    assert!(result.is_ok(), "{:}", result.err().unwrap());
+}
 
 // TEST UTILS
 // ================================================================================================
@@ -23,13 +57,10 @@ pub fn build_prover_channel(
     DefaultProverChannel::new(trace_length * options.blowup_factor(), 32)
 }
 
-pub fn build_evaluations(
-    trace_length: usize,
-    lde_blowup: usize,
-    ce_blowup: usize,
-) -> Vec<BaseElement> {
-    let len = (trace_length * ce_blowup) as u128;
-    let mut p = (0..len).map(BaseElement::new).collect::<Vec<_>>();
+pub fn build_evaluations(trace_length: usize, lde_blowup: usize) -> Vec<BaseElement> {
+    let mut p = (0..trace_length as u128)
+        .map(BaseElement::new)
+        .collect::<Vec<_>>();
     let domain_size = trace_length * lde_blowup;
     p.resize(domain_size, BaseElement::ZERO);
 
@@ -71,13 +102,13 @@ pub fn verify_proof(
         })
         .collect::<Vec<BaseElement>>();
     let context = VerifierContext::<BaseElement, BaseElement, Blake3>::new(
-        evaluations.len(),
         max_degree,
         commitments,
         alphas,
         channel.num_partitions(),
         options.clone(),
     );
+    assert_eq!(evaluations.len(), context.domain_size());
     let queried_evaluations = positions
         .iter()
         .map(|&p| evaluations[p])
