+++ fri/src/verifier/channel.rs
@@ -3,46 +3,95 @@
 // This source code is licensed under the MIT license found in the
 // LICENSE file in the root directory of this source tree.
 
-use crate::{utils::hash_values, FriProof, ProofSerializationError, VerifierError};
+use crate::{utils::hash_values, FriProof, VerifierError};
 use crypto::{BatchMerkleProof, ElementHasher, Hasher, MerkleTree};
 use math::FieldElement;
-use utils::{group_vector_elements, transpose_slice};
+use utils::{group_vector_elements, transpose_slice, DeserializationError};
 
 // VERIFIER CHANNEL TRAIT
 // ================================================================================================
 
+/// Defines an abstraction for reading data sent by the prover during query phase of the FRI
+/// protocol.
+///
+/// This trait abstracts away implementation specifics of the [FriProof] struct. Thus, instead of
+/// dealing with FRI proofs directly, the verifier can read the data as if it was sent by the
+/// prover via an interactive channel.
+///
+/// The implementors of this trait should provide implementations for `take*()` methods, while the
+/// users of the trait, should call `read*()` methods to read the desired data. Note, that reading
+/// removes the read data from the channel. Thus, reading duplicated values from the channel
+/// should not be possible.
 pub trait VerifierChannel<E: FieldElement> {
+    /// Hash function used by the prover to commit to polynomial evaluations.
     type Hasher: ElementHasher<BaseField = E::BaseField>;
 
     // REQUIRED METHODS
     // --------------------------------------------------------------------------------------------
 
-    fn take_next_fri_layer_proof(&mut self) -> BatchMerkleProof<Self::Hasher>;
+    /// Reads and removes from the channel evaluations of the polynomial at the queried positions
+    /// for the next FRI layer.
+    ///
+    /// In the interactive version of the protocol, these evaluations are sent from the prover to
+    /// the verifier during the query phase of the FRI protocol.
+    ///
+    /// It is expected that layer queries and layer proofs at the same FRI layer are consistent.
+    /// That is, query values hash into the leaf nodes of corresponding Merkle authentication
+    /// paths.
     fn take_next_fri_layer_queries(&mut self) -> Vec<E>;
+
+    /// Reads and removes from the channel Merkle authentication paths for queried evaluations for
+    /// the next FRI layer.
+    ///
+    /// In the interactive version of the protocol, these authentication paths are sent from the
+    /// prover to the verifier during the query phase of the FRI protocol.
+    ///
+    /// It is expected that layer proofs and layer queries at the same FRI layer are consistent.
+    /// That is, query values hash into the leaf nodes of corresponding Merkle authentication
+    /// paths.
+    fn take_next_fri_layer_proof(&mut self) -> BatchMerkleProof<Self::Hasher>;
+
+    /// Reads and removes the remainder (last FRI layer) values from the channel.
+    ///
+    /// In the interactive version of the protocol, the remainder is sent from the prover to the
+    /// verifier during the query phase of the FRI protocol.
     fn take_fri_remainder(&mut self) -> Vec<E>;
 
     // PROVIDED METHODS
     // --------------------------------------------------------------------------------------------
 
-    /// Returns FRI query values at the specified positions from the FRI layer at the
-    /// specified index. This also checks if the values are valid against the provided
-    /// FRI layer commitment.
+    /// Returns FRI query values at the specified positions from the current FRI layer and advances
+    /// layer pointer by one.
+    ///
+    /// This also checks if the values are valid against the provided FRI layer commitment.
+    ///
+    /// # Errors
+    /// Returns an error if query values did not match layer commitment.
     fn read_layer_queries<const N: usize>(
         &mut self,
-        layer_idx: usize,
         positions: &[usize],
         commitment: &<<Self as VerifierChannel<E>>::Hasher as Hasher>::Digest,
     ) -> Result<Vec<[E; N]>, VerifierError> {
         let layer_proof = self.take_next_fri_layer_proof();
         MerkleTree::<Self::Hasher>::verify_batch(commitment, positions, &layer_proof)
-            .map_err(|_| VerifierError::LayerCommitmentMismatch(layer_idx))?;
+            .map_err(|_| VerifierError::LayerCommitmentMismatch)?;
+
+        // TODO: make sure layer queries hash into leaves of layer proof
 
         let layer_queries = self.take_next_fri_layer_queries();
         Ok(group_vector_elements(layer_queries))
     }
 
-    /// Reads FRI remainder values (last FRI layer). This also checks that the remainder is
-    /// valid against the provided commitment.
+    /// Returns FRI remainder values (last FRI layer) read from this channel.
+    ///
+    /// This also checks whether the remainder is valid against the provided commitment.
+    ///
+    /// # Errors
+    /// Returns an error if:
+    /// - Remainder values read from the channel cannot be used to construct a fully-balanced
+    ///   Merkle tree.
+    /// - If the root of the Merkle tree constructed from the remainder values does not match
+    ///   the specified `commitment`.
     fn read_remainder<const N: usize>(
         &mut self,
         commitment: &<<Self as VerifierChannel<E>>::Hasher as Hasher>::Digest,
@@ -53,7 +102,7 @@ pub trait VerifierChannel<E: FieldElement> {
         let remainder_values = transpose_slice(&remainder);
         let hashed_values = hash_values::<Self::Hasher, E, N>(&remainder_values);
         let remainder_tree = MerkleTree::<Self::Hasher>::new(hashed_values)
-            .expect("failed to construct FRI remainder tree");
+            .map_err(|err| VerifierError::RemainderTreeConstructionFailed(format!("{}", err)))?;
 
         // make sure the root of the tree matches the committed root of the last layer
         if commitment != remainder_tree.root() {
@@ -67,6 +116,12 @@ pub trait VerifierChannel<E: FieldElement> {
 // DEFAULT VERIFIER CHANNEL IMPLEMENTATION
 // ================================================================================================
 
+/// Provides a default implementation of the [VerifierChannel] trait.
+///
+/// Default verifier channel can be instantiated directly from a [FriProof] struct.
+///
+/// Though this implementation is primarily intended for testing purposes, it can be used in
+/// production use cases as well.
 pub struct DefaultVerifierChannel<E: FieldElement, H: ElementHasher<BaseField = E::BaseField>> {
     layer_proofs: Vec<BatchMerkleProof<H>>,
     layer_queries: Vec<Vec<E>>,
@@ -79,12 +134,15 @@ where
     E: FieldElement,
     H: ElementHasher<BaseField = E::BaseField>,
 {
-    /// Builds a new verifier channel from the specified parameters.
+    /// Builds a new verifier channel from the specified [FirProof].
+    ///
+    /// # Errors
+    /// Returns an error if the specified `proof` could not be parsed correctly.
     pub fn new(
         proof: FriProof,
         domain_size: usize,
         folding_factor: usize,
-    ) -> Result<Self, ProofSerializationError> {
+    ) -> Result<Self, DeserializationError> {
         let num_partitions = proof.num_partitions();
 
         let remainder = proof.parse_remainder()?;
@@ -99,6 +157,7 @@ where
         })
     }
 
+    /// Returns the number of partitions used during proof generation.
     pub fn num_partitions(&self) -> usize {
         self.num_partitions
     }
